#2670: move reportError to telemetry directory
