+++ src/actionPanel/protocol.tsx
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { reportError } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import { FormEntry, PanelEntry } from "@/actionPanel/actionPanelTypes";
 import { FormDefinition } from "@/blocks/transformers/ephemeralForm/formTypes";
 import { UUID } from "@/core";
