+++ src/auth/ScopeSettings.tsx
@@ -24,7 +24,8 @@ import { faEyeSlash, faInfo } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { StatusCodes } from "http-status-codes";
 import { getLinkedApiClient } from "@/services/apiClient";
-import { isAxiosError, reportError } from "@/errors";
+import { isAxiosError } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import useNotifications from "@/hooks/useNotifications";
 import ConnectedFieldTemplate from "@/components/form/ConnectedFieldTemplate";
 import styles from "./ScopeSettings.module.scss";
