+++ src/background/contextMenus.ts
@@ -18,7 +18,8 @@
 import pTimeout from "p-timeout";
 import browser, { Menus, Tabs } from "webextension-polyfill";
 import { isBackground } from "webext-detect-page";
-import { getErrorMessage, hasCancelRootCause, reportError } from "@/errors";
+import { getErrorMessage, hasCancelRootCause } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import { noop } from "lodash";
 import {
   handleMenuAction,
