+++ src/devTools/context.ts
@@ -21,7 +21,8 @@ import browser from "webextension-polyfill";
 import { navigationEvent } from "@/devTools/events";
 import { useAsyncEffect } from "use-async-effect";
 import { FrameworkMeta } from "@/messaging/constants";
-import { getErrorMessage, isErrorObject, reportError } from "@/errors";
+import { getErrorMessage, isErrorObject } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import { uuidv4 } from "@/types/helpers";
 import { useTabEventListener } from "@/hooks/events";
 import { thisTab } from "@/devTools/utils";
