+++ src/devTools/editor/fields/ConfigErrorBoundary.tsx
@@ -16,7 +16,8 @@
  */
 
 import React, { Component } from "react";
-import { getErrorMessage, reportError } from "@/errors";
+import { getErrorMessage } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import { UnknownObject } from "@/types";
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
