+++ src/devTools/editor/hooks/useAddElement.ts
@@ -22,7 +22,8 @@ import { useGetAuthQuery } from "@/services/api";
 import { useToasts } from "react-toast-notifications";
 import { actions, FormState } from "@/devTools/editor/slices/editorSlice";
 import { internalExtensionPointMetaFactory } from "@/devTools/editor/extensionPoints/base";
-import { getErrorMessage, reportError } from "@/errors";
+import { getErrorMessage } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import { ElementConfig } from "@/devTools/editor/extensionPoints/elementConfig";
 import { getCurrentURL, thisTab } from "@/devTools/utils";
 import { updateDynamicElement } from "@/contentScript/messenger/api";
