+++ src/devTools/editor/hooks/useRemove.ts
@@ -21,7 +21,8 @@ import { useToasts } from "react-toast-notifications";
 import { useFormikContext } from "formik";
 import { useDispatch } from "react-redux";
 import { useModals } from "@/components/ConfirmationModal";
-import { getErrorMessage, reportError } from "@/errors";
+import { getErrorMessage } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import { uninstallContextMenu } from "@/background/messenger/api";
 import { thisTab } from "@/devTools/utils";
 import {
