+++ src/hooks/useBlockOptions.ts
@@ -22,7 +22,7 @@ import genericOptionsFactory, {
 import { IBlock, RegistryId } from "@/core";
 import blockRegistry from "@/blocks/registry";
 import { useAsyncEffect } from "use-async-effect";
-import { reportError } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import optionsRegistry from "@/components/fields/optionsRegistry";
 
 interface BlockState {
