+++ src/popups/PermissionsPopup.tsx
@@ -21,7 +21,8 @@ import { faShieldAlt } from "@fortawesome/free-solid-svg-icons";
 import AsyncButton from "@/components/AsyncButton";
 import { Button } from "react-bootstrap";
 import Centered from "@/devTools/editor/components/Centered";
-import { getErrorMessage, reportError } from "@/errors";
+import { getErrorMessage } from "@/errors";
+import reportError from "@/telemetry/reportError";
 import browser, { Permissions } from "webextension-polyfill";
 import { selectOptionalPermissions } from "@/utils/permissions";
