+++ src/telemetry/reportUncaughtErrors.ts
@@ -19,7 +19,8 @@
  * @file This file must be imported as early as possible in each entrypoint, once
  */
 import { isBackground } from "webext-detect-page";
-import { getErrorMessage, IGNORED_ERRORS, reportError } from "@/errors";
+import { getErrorMessage, IGNORED_ERRORS } from "@/errors";
+import reportError from "@/telemetry/reportError";
 
 function ignoreSomeErrors(
   errorEvent: ErrorEvent | PromiseRejectionEvent
